+++ src/main/java/uk/gov/hmcts/darts/event/client/EventApiClient.java
@@ -1,10 +1,12 @@
-package uk.gov.hmcts.darts.events;
+package uk.gov.hmcts.darts.event.client;
 
 import feign.Headers;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import uk.gov.hmcts.darts.config.FeignConfig;
+import uk.gov.hmcts.darts.event.model.EventRequest;
+import uk.gov.hmcts.darts.event.model.EventResponse;
 
 import static org.springframework.web.bind.annotation.RequestMethod.POST;
