+++ src/main/java/uk/gov/hmcts/darts/event/model/EventResponse.java
@@ -1,7 +1,8 @@
-package uk.gov.hmcts.darts.events;
+package uk.gov.hmcts.darts.event.model;
 
 // This will be moved to the API repo or a library
 public record EventResponse(
     String message,
     String code) {
+
 }
