+++ src/main/java/uk/gov/hmcts/darts/event/service/impl/DarNotifyEventServiceImpl.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.darts.events.service.impl;
+package uk.gov.hmcts.darts.event.service.impl;
 
 import com.viqsoultions.DARNotifyEvent;
 import com.viqsoultions.Event;
@@ -8,10 +8,10 @@ import com.viqsoultions.XMLEventDocument;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
-import uk.gov.hmcts.darts.events.client.DarNotifyEventClient;
-import uk.gov.hmcts.darts.events.config.DarNotifyEventConfigurationProperties;
-import uk.gov.hmcts.darts.events.model.DarNotifyEvent;
-import uk.gov.hmcts.darts.events.service.DarNotifyEventService;
+import uk.gov.hmcts.darts.event.client.DarNotifyEventClient;
+import uk.gov.hmcts.darts.event.config.DarNotifyEventConfigurationProperties;
+import uk.gov.hmcts.darts.event.model.DarNotifyEvent;
+import uk.gov.hmcts.darts.event.service.DarNotifyEventService;
 
 import java.time.LocalDateTime;
 import java.time.OffsetDateTime;
