+++ src/main/java/uk/gov/hmcts/darts/event/service/impl/EventRoute.java
@@ -1,10 +1,12 @@
-package uk.gov.hmcts.darts.events;
+package uk.gov.hmcts.darts.event.service.impl;
 
 import com.service.mojdarts.synapps.com.AddDocumentResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import uk.gov.courtservice.events.DartsEvent;
+import uk.gov.hmcts.darts.event.client.EventApiClient;
+import uk.gov.hmcts.darts.event.model.EventResponse;
 import uk.gov.hmcts.darts.utilities.XmlParser;
 import uk.gov.hmcts.darts.utilities.XmlValidator;
